+++ src/main/kotlin/com/wheretopop/domain/building/core/BuildingCriteria.kt
@@ -1,4 +1,4 @@
-package com.wheretopop.domain.building
+package com.wheretopop.domain.building.core
 
 
 /**
