+++ src/main/kotlin/com/wheretopop/domain/building/core/BuildingInfoMapper.kt
@@ -1,4 +1,4 @@
-package com.wheretopop.domain.building
+package com.wheretopop.domain.building.core
 
 /**
  * 도메인 객체에서 DTO로 변환하는 매퍼 클래스
