+++ src/main/kotlin/com/wheretopop/domain/building/core/BuildingService.kt
@@ -1,4 +1,4 @@
-package com.wheretopop.domain.building
+package com.wheretopop.domain.building.core
 
 interface BuildingService {
     suspend fun searchBuildings(criteria: BuildingCriteria.SearchBuildingCriteria): List<BuildingInfo.Main>
