+++ src/main/kotlin/com/wheretopop/domain/building/register/BuildingRegisterServiceImpl.kt
@@ -1,7 +1,5 @@
 package com.wheretopop.domain.building.register
 
-import com.wheretopop.domain.building.BuildingCriteria
-import com.wheretopop.domain.building.BuildingInfo
 import org.springframework.stereotype.Service
 import java.time.Instant
